moved user.dart to models
as it should sksk
